[i-27] moving monkeypatches to their folders
